# iap-api/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication Controller"
                ],
                "summary": "Login a user",
                "description": "Login a user and return access and refresh tokens",
                "operationId": "exchangeCodeForTokens",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TokensDTO"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token exchange failed"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "LoginDTO": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string"
                    }
                }
            },
            "TokensDTO": {
                "type": "object",
                "properties": {
                    "accessToken": {
                        "type": "string"
                    },
                    "refreshToken": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}